Add tests for PI selection eligiblity in project request form; refactor test logic
