+++ coldfront/core/project/tests/test_forms/test_new_project_forms/test_savio_project_existing_pi_form.py
@@ -1,7 +1,7 @@
 from coldfront.core.allocation.models import AllocationPeriod
 from coldfront.core.allocation.models import AllocationRenewalRequest
 from coldfront.core.allocation.models import AllocationRenewalRequestStatusChoice
-from coldfront.core.project.forms_.renewal_forms.request_forms import ProjectRenewalPISelectionForm
+from coldfront.core.project.forms_.new_project_forms.request_forms import SavioProjectExistingPIForm
 from coldfront.core.project.models import Project
 from coldfront.core.project.models import ProjectAllocationRequestStatusChoice
 from coldfront.core.project.models import ProjectStatusChoice
@@ -15,8 +15,8 @@ from coldfront.core.utils.common import utc_now_offset_aware
 from coldfront.core.utils.tests.test_base import TestBase
 
 
-class TestProjectRenewalPISelectionForm(TestBase):
-    """A class for testing ProjectRenewalPISelectionForm."""
+class TestSavioProjectExistingPIForm(TestBase):
+    """A class for testing SavioProjectExistingPIForm."""
 
     def setUp(self):
         """Set up test data."""
@@ -25,23 +25,6 @@ class TestProjectRenewalPISelectionForm(TestBase):
         self.sign_user_access_agreement(self.user)
         self.client.login(username=self.user.username, password=self.password)
 
-        # Create a Project for the user to renew.
-        project_name = 'fc_project'
-        active_project_status = ProjectStatusChoice.objects.get(name='Active')
-        self.project = Project.objects.create(
-            name=project_name,
-            title=project_name,
-            status=active_project_status)
-        pi_role = ProjectUserRoleChoice.objects.get(
-            name='Principal Investigator')
-        active_project_user_status = ProjectUserStatusChoice.objects.get(
-            name='Active')
-        self.project_user = ProjectUser.objects.create(
-            project=self.project,
-            role=pi_role,
-            status=active_project_user_status,
-            user=self.user)
-
     def test_eligibility_based_on_requests_in_specific_allocation_period(self):
         """Test that PI eligibility for a particular AllocationPeriod is
         only based on existing requests under the same period."""
@@ -49,11 +32,29 @@ class TestProjectRenewalPISelectionForm(TestBase):
 
         # Create a new project request.
         new_project, new_project_request = create_fca_project_and_request(
-            'fc_new_project', 'New', allocation_period, self.user, self.user,
-            'Under Review')
-        self.assertNotEqual(new_project_request.status.name, 'Denied')
+            'fc_new_project', 'Denied', allocation_period, self.user,
+            self.user, 'Under Review')
 
-        # Create a renewal request.
+        # Create a Project for the user to renew.
+        project_name = 'fc_project'
+        # Use 'Denied' for testing since another check disables PIs with
+        # Projects having the other statuses.
+        denied_project_status = ProjectStatusChoice.objects.get(name='Denied')
+        project = Project.objects.create(
+            name=project_name,
+            title=project_name,
+            status=denied_project_status)
+        pi_role = ProjectUserRoleChoice.objects.get(
+            name='Principal Investigator')
+        active_project_user_status = ProjectUserStatusChoice.objects.get(
+            name='Active')
+        ProjectUser.objects.create(
+            project=project,
+            role=pi_role,
+            status=active_project_user_status,
+            user=self.user)
+
+        # Create an AllocationRenewalRequest.
         under_review_request_status = \
             AllocationRenewalRequestStatusChoice.objects.get(
                 name='Under Review')
@@ -62,10 +63,9 @@ class TestProjectRenewalPISelectionForm(TestBase):
             pi=self.user,
             allocation_period=allocation_period,
             status=under_review_request_status,
-            pre_project=self.project,
-            post_project=self.project,
+            pre_project=project,
+            post_project=project,
             request_time=utc_now_offset_aware())
-        self.assertNotEqual(allocation_renewal_request.status.name, 'Denied')
 
         # Select a different AllocationPeriod.
         next_allowance_year_allocation_period = \
@@ -74,15 +74,15 @@ class TestProjectRenewalPISelectionForm(TestBase):
                 start_date__gt=display_time_zone_current_date()).first()
         self.assertIsNotNone(next_allowance_year_allocation_period)
         kwargs = {
-            'allocation_period_pk': next_allowance_year_allocation_period.pk,
-            'project_pks': [self.project.pk],
+            'allocation_type': 'FCA',
+            'allocation_period': next_allowance_year_allocation_period,
         }
 
         # The PI should be selectable.
-        form = ProjectRenewalPISelectionForm(**kwargs)
+        form = SavioProjectExistingPIForm(**kwargs)
         pi_field_disabled_choices = \
             form.fields['PI'].widget.disabled_choices
-        self.assertNotIn(self.project_user.pk, pi_field_disabled_choices)
+        self.assertNotIn(self.user.pk, pi_field_disabled_choices)
 
         # Change the AllocationPeriods of the requests.
         new_project_request.allocation_period = \
@@ -93,10 +93,34 @@ class TestProjectRenewalPISelectionForm(TestBase):
         allocation_renewal_request.save()
 
         # The PI should not be selectable.
-        form = ProjectRenewalPISelectionForm(**kwargs)
+        form = SavioProjectExistingPIForm(**kwargs)
         pi_field_disabled_choices = \
             form.fields['PI'].widget.disabled_choices
-        self.assertIn(self.project_user.pk, pi_field_disabled_choices)
+        self.assertIn(self.user.pk, pi_field_disabled_choices)
+
+    def test_pis_with_inactive_fc_projects_disabled(self):
+        """Test that PIs of Projects with the 'Inactive' status are
+        disabled in the 'PI' field."""
+        inactive_name = 'fc_inactive_project'
+        inactive_status = ProjectStatusChoice.objects.get(name='Inactive')
+        inactive_project = Project.objects.create(
+            name=inactive_name, title=inactive_name, status=inactive_status)
+
+        # Add the user as a PI on both Projects.
+        pi_role = ProjectUserRoleChoice.objects.get(
+            name='Principal Investigator')
+        active_status = ProjectUserStatusChoice.objects.get(name='Active')
+        kwargs = {
+            'project': inactive_project,
+            'role': pi_role,
+            'status': active_status,
+            'user': self.user,
+        }
+        ProjectUser.objects.create(**kwargs)
+
+        form = SavioProjectExistingPIForm(allocation_type='FCA')
+        pi_field_disabled_choices = form.fields['PI'].widget.disabled_choices
+        self.assertIn(self.user.pk, pi_field_disabled_choices)
 
     def test_pis_with_non_denied_project_allocation_requests_disabled(self):
         """Test that PIs with non-'Denied'
@@ -105,34 +129,51 @@ class TestProjectRenewalPISelectionForm(TestBase):
         allocation_period = get_current_allowance_year_period()
         # Create a new project request.
         new_project, new_project_request = create_fca_project_and_request(
-            'fc_new_project', 'New', allocation_period, self.user, self.user,
-            'Under Review')
+            'fc_new_project', 'Denied', allocation_period, self.user,
+            self.user, 'Under Review')
 
         # For every status except 'Denied', the PI should be disabled.
         kwargs = {
-            'allocation_period_pk': allocation_period.pk,
-            'project_pks': [self.project.pk],
+            'allocation_type': 'FCA',
+            'allocation_period': allocation_period,
         }
         status_choices = ProjectAllocationRequestStatusChoice.objects.all()
         self.assertEqual(status_choices.count(), 5)
         for status_choice in status_choices:
             new_project_request.status = status_choice
             new_project_request.save()
-            form = ProjectRenewalPISelectionForm(**kwargs)
+            form = SavioProjectExistingPIForm(**kwargs)
             pi_field_disabled_choices = \
                 form.fields['PI'].widget.disabled_choices
             if status_choice.name != 'Denied':
-                self.assertIn(
-                    self.project_user.pk, pi_field_disabled_choices)
+                self.assertIn(self.user.pk, pi_field_disabled_choices)
             else:
-                self.assertNotIn(
-                    self.project_user.pk, pi_field_disabled_choices)
+                self.assertNotIn(self.user.pk, pi_field_disabled_choices)
 
     def test_pis_with_non_denied_allocation_renewal_requests_disabled(self):
-        """Test that PIs with non-'Denied' AllocationRenewalRequests are
-        disabled in the 'PI' field."""
+        """Test that PIs with non-'Denied' AllocationRenewalRequests
+        are disabled in the 'PI' field."""
+        # Create a Project for the user to renew.
+        project_name = 'fc_project'
+        # Use 'Denied' for testing since another check disables PIs with
+        # Projects having the other statuses.
+        denied_project_status = ProjectStatusChoice.objects.get(name='Denied')
+        project = Project.objects.create(
+            name=project_name,
+            title=project_name,
+            status=denied_project_status)
+        pi_role = ProjectUserRoleChoice.objects.get(
+            name='Principal Investigator')
+        active_project_user_status = ProjectUserStatusChoice.objects.get(
+            name='Active')
+        ProjectUser.objects.create(
+            project=project,
+            role=pi_role,
+            status=active_project_user_status,
+            user=self.user)
+
         allocation_period = get_current_allowance_year_period()
-        # Create a renewal request.
+        # Create an AllocationRenewalRequest.
         under_review_request_status = \
             AllocationRenewalRequestStatusChoice.objects.get(
                 name='Under Review')
@@ -141,26 +182,27 @@ class TestProjectRenewalPISelectionForm(TestBase):
             pi=self.user,
             allocation_period=allocation_period,
             status=under_review_request_status,
-            pre_project=self.project,
-            post_project=self.project,
+            pre_project=project,
+            post_project=project,
             request_time=utc_now_offset_aware())
 
         # For every status except 'Denied', the PI should be disabled.
         kwargs = {
-            'allocation_period_pk': allocation_period.pk,
-            'project_pks': [self.project.pk],
+            'allocation_type': 'FCA',
+            'allocation_period': allocation_period,
         }
         status_choices = AllocationRenewalRequestStatusChoice.objects.all()
         self.assertEqual(status_choices.count(), 4)
         for status_choice in status_choices:
             allocation_renewal_request.status = status_choice
             allocation_renewal_request.save()
-            form = ProjectRenewalPISelectionForm(**kwargs)
+            form = SavioProjectExistingPIForm(**kwargs)
+
             pi_field_disabled_choices = \
                 form.fields['PI'].widget.disabled_choices
             if status_choice.name != 'Denied':
-                self.assertIn(
-                    self.project_user.pk, pi_field_disabled_choices)
+                self.assertIn(self.user.pk, pi_field_disabled_choices)
             else:
-                self.assertNotIn(
-                    self.project_user.pk, pi_field_disabled_choices)
+                self.assertNotIn(self.user.pk, pi_field_disabled_choices)
+
+    # TODO
